refactor: make vsss module local
